# docs/swagger.yaml
definitions:
  http.CreateUserRequest:
    properties:
      age:
        maximum: 150
        minimum: 0
        type: integer
      email:
        type: string
      firstname:
        maxLength: 50
        minLength: 2
        type: string
      lastname:
        maxLength: 50
        minLength: 2
        type: string
      phone:
        type: string
    required:
    - email
    - firstname
    - lastname
    type: object
  http.UserRequest:
    properties:
      age:
        maximum: 150
        minimum: 0
        type: integer
      email:
        type: string
      firstname:
        maxLength: 50
        minLength: 2
        type: string
      lastname:
        maxLength: 50
        minLength: 2
        type: string
      phone:
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    type: object
  http.UserResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      phone:
        type: string
      status:
        type: string
      userId:
        type: string
    type: object
info:
  contact: {}
  description: This api allow to create, modify,delete, and retrieve user records.
  title: User Management API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Retrieves all users from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a user with first name, last name, and email, and other
        optional data.
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.UserResponse'
      summary: Create a new user
      tags:
      - users
  /users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete an existing user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieves all users from the database.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update a user with first name, last name, and email.
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/http.UserRequest'
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.UserResponse'
      summary: Update an existing user
      tags:
      - users
swagger: "2.0"